requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1